chore: Refactor `rover new` into `rover template`
