+++ src/cli.rs
@@ -198,7 +198,7 @@ impl Rover {
                 self.get_checks_timeout_seconds()?,
                 self.get_json(),
             ),
-            Command::New(command) => command.run(self.get_client_config()?),
+            Command::Template(command) => command.run(self.get_client_config()?),
             Command::Readme(command) => command.run(self.get_client_config()?),
             Command::Subgraph(command) => command.run(
                 self.get_client_config()?,
@@ -360,8 +360,8 @@ pub enum Command {
     /// Graph API schema commands
     Graph(command::Graph),
 
-    /// New template commands
-    New(command::New),
+    /// Commands for working with templates
+    Template(command::Template),
 
     /// Readme commands
     Readme(command::Readme),
