+++ src/command/mod.rs
@@ -6,10 +6,10 @@ mod fed2;
 mod graph;
 mod info;
 pub(crate) mod install;
-mod new;
 mod readme;
 pub(crate) mod subgraph;
 mod supergraph;
+mod template;
 mod update;
 
 pub(crate) mod output;
@@ -22,9 +22,9 @@ pub use fed2::Fed2;
 pub use graph::Graph;
 pub use info::Info;
 pub use install::Install;
-pub use new::New;
 pub use output::RoverOutput;
 pub use readme::Readme;
 pub use subgraph::Subgraph;
 pub use supergraph::Supergraph;
+pub use template::Template;
 pub use update::Update;
